move VideoSettings to cores
