+++ xbmc/cores/VideoPlayer/DVDCodecs/DVDFactoryCodec.cpp
@@ -43,7 +43,7 @@
 #include "DVDStreamInfo.h"
 #include "settings/AdvancedSettings.h"
 #include "settings/Settings.h"
-#include "settings/VideoSettings.h"
+#include "cores/VideoSettings.h"
 #include "threads/SingleLock.h"
 #include "utils/StringUtils.h"
