+++ xbmc/cores/VideoSettings.cpp
@@ -17,15 +17,8 @@
  *  <http://www.gnu.org/licenses/>.
  *
  */
-// VideoSettings.cpp: implementation of the CVideoSettings class.
-//
-//////////////////////////////////////////////////////////////////////
 
 #include "VideoSettings.h"
-
-//////////////////////////////////////////////////////////////////////
-// Construction/Destruction
-//////////////////////////////////////////////////////////////////////
 
 CVideoSettings::CVideoSettings()
 {
