+++ xbmc/settings/MediaSettings.h
@@ -26,7 +26,7 @@
 #include "settings/lib/ISettingsHandler.h"
 #include "settings/lib/ISubSettings.h"
 #include "settings/GameSettings.h"
-#include "settings/VideoSettings.h"
+#include "cores/VideoSettings.h"
 #include "threads/CriticalSection.h"
 
 #define VOLUME_DRC_MINIMUM 0    // 0dB
